app/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6  # Usar psycopg2-binary para evitar problemas con pg_config

